requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
peft>=0.10.0
trl>=0.8.0

# Vision & Video
decord>=0.6.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
qwen-vl-utils>=0.0.2

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data Pipeline
openpack-toolkit>=1.1.0
webdataset>=0.2.86
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Training
deepspeed>=0.14.0
wandb>=0.16.0

# Evaluation
scikit-learn>=1.3.0
tqdm>=4.65.0

# Utils
pyyaml>=6.0
